+++ web_page/main.html
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="style.css">
+    <title>Poznan Theft</title>
+</head>
+<body>
+    <header>
+        <div class="header">
+
+        </div>
+
+    </header>
+    <div class="main">
+        <div class="add_menu">
+            <div class="text">
+
+            </div>
+            <div class="window">
+                <input type="text" placeholder="Type address"  name="" ,value="">
+
+            </div>
+            <div class="button">
+                <button type="submit"></button>
+
+            </div>
+        </div>
+
+        <div class="search_menu">
+            <div class="text">
+                ""
+            </div>
+            <div class="window">
+                <input type="text">
+            </div>
+            <div class="button">
+
+            </div>
+        </div>
+
+    </div>
+
+    <div class="footer">
+
+
+    </div>
+    
+</body>
+</html>
